{"schema":"decodex/commit/1","summary":"Split MCP observability sanitizer modules","authority":"XY-1115"}
