+++ apps/decodex/src/mcp/observability/sanitizer/reasoning.rs
@@ -1,0 +1,35 @@
+use serde_json::Value;
+
+pub(super) fn redact_reasoning_protocol_activity(value: &mut Value) {
+	match value {
+		Value::Object(object) => {
+			let is_reasoning_event = object
+				.get("category")
+				.and_then(Value::as_str)
+				.is_some_and(|category| category.eq_ignore_ascii_case("reasoning"))
+				|| object.get("event_type").and_then(Value::as_str).is_some_and(|event_type| {
+					event_type.to_ascii_lowercase().contains("reasoning")
+				});
+
+			if is_reasoning_event {
+				object.insert(
+					String::from("detail"),
+					Value::String(String::from("redacted_reasoning")),
+				);
+				object.remove("text");
+				object.remove("summary");
+				object.remove("content");
+				object.remove("body");
+			}
+
+			for child in object.values_mut() {
+				redact_reasoning_protocol_activity(child);
+			}
+		},
+		Value::Array(items) =>
+			for item in items {
+				redact_reasoning_protocol_activity(item);
+			},
+		_ => {},
+	}
+}
